requirements.txt
Flask
readability-api
python-instagram
httplib2